refactored jacobbians for GLMs
