+++ batchglm/train/tf/base_glm_all/external.py
@@ -4,7 +4,7 @@ import batchglm.train.tf.ops as op_utils
 import batchglm.train.tf.train as train_utils
 from batchglm.train.tf.base import TFEstimatorGraph, MonitoredTFEstimator
 from batchglm.train.tf.base_glm import GradientGraphGLM, NewtonGraphGLM, TrainerGraphGLM, EstimatorGraphGLM, FullDataModelGraphGLM, BasicModelGraphGLM
-from batchglm.train.tf.base_glm import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM, HessiansGLM
+from batchglm.train.tf.base_glm import ESTIMATOR_PARAMS, ProcessModelGLM, ModelVarsGLM, FIMGLM, HessiansGLM, JacobiansGLM
 
 from batchglm.models.base_glm import InputData, _Model_GLM
